+++ app/src/main/java/app/milanherke/mystudiez/AddEditSubjectActivity.kt
@@ -14,6 +14,9 @@ import androidx.lifecycle.ViewModelProviders
 import app.milanherke.mystudiez.ActivityUtils.Companion.ACTIVITY_NAME_BUNDLE_ID
 import app.milanherke.mystudiez.ActivityUtils.Companion.FRAGMENT_TO_LOAD_BUNDLE_ID
 import app.milanherke.mystudiez.ActivityUtils.Companion.SUBJECT_PARAM_BUNDLE_ID
+import app.milanherke.mystudiez.fragments.SubjectDetailsFragment
+import app.milanherke.mystudiez.fragments.SubjectsFragment
+import app.milanherke.mystudiez.fragments.UnsavedChangesDialogFragment
 import kotlinx.android.synthetic.main.activity_add_edit_subject.*
 import kotlinx.android.synthetic.main.content_add_edit_subject.*
