+++ app/src/main/java/app/milanherke/mystudiez/fragments/UnsavedChangesDialogFragment.kt
@@ -1,9 +1,10 @@
-package app.milanherke.mystudiez
+package app.milanherke.mystudiez.fragments
 
 import android.app.Dialog
 import android.os.Bundle
 import androidx.appcompat.app.AlertDialog
 import androidx.fragment.app.DialogFragment
+import app.milanherke.mystudiez.R
 
 class UnsavedChangesDialogFragment(private val listener: DialogInteractions) : DialogFragment() {
 
@@ -20,7 +21,7 @@ class UnsavedChangesDialogFragment(private val listener: DialogInteractions) : D
                 .setPositiveButton(R.string.changes_dialog_continue) { _, _ ->
                     listener.onPositiveBtnPressed()
                 }
-                .setNegativeButton(R.string.changes_dialog_back) {_, _ ->
+                .setNegativeButton(R.string.changes_dialog_back) { _, _ ->
                     listener.onNegativeBtnPressed()
                     dismiss()
                 }
