+++ app/src/main/res/layout/fragment_exam_details.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:orientation="vertical"
-    tools:context=".ExamDetailsFragment">
+    tools:context=".fragments.ExamDetailsFragment">
 
     <androidx.cardview.widget.CardView
         android:id="@+id/exam_details_cardview"
