+++ app/src/main/res/layout/fragment_lesson_details.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:orientation="vertical"
-    tools:context=".LessonDetailsFragment">
+    tools:context=".fragments.LessonDetailsFragment">
 
     <androidx.cardview.widget.CardView
         android:id="@+id/lesson_details_cardview"
