+++ app/src/main/res/layout/fragment_task_details.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:orientation="vertical"
-    tools:context=".TaskDetailsFragment">
+    tools:context=".fragments.TaskDetailsFragment">
 
     <androidx.cardview.widget.CardView
         android:id="@+id/task_details_cardview"
