+++ app/src/main/res/layout/fragment_tasks.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".TasksFragment">
+    tools:context=".fragments.TasksFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/task_list"
